requirements.txt
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
pydantic>=2.11.7
sqlalchemy>=2.0.43
requests>=2.32.5

# LLM / LangChain stack (optional)
langchain>=0.3.27
langchain-ollama>=0.3.7
langchain-community>=0.3.29
chromadb>=1.0.20

# Numerical & plotting
numpy>=2.3.2
scipy>=1.11.0
pandas>=2.2.3
matplotlib>=3.8.0
nashpy>=0.0.41

# UI
streamlit>=1.30.0
plotly>=6.0.0
python-dotenv>=1.0.1
